Level insertion implemented
- Level manager class created for spawning and loading levels and
layouts
- Grid converted to a level class, for holding the state of the current
environment
- Tile class now uses inheritence so that floors, walls and other blocks
have the same basic functionality
- The same mesh manager is passed through all these objects for later
use
- An ENUM has been created for the different types of blocks. Paired
with a switch statement, this can be used for generating children and
retrieving models quickly as well as be looked at by other classes.
- Levels have layers - a floor and a body. The floor is for the blocks
you walk on, and can be switched out for special floors, and the body is
for walls, pickups, spawns and doors
- The level class has been designed to be re-initialised. This needs
some testing as there's no functionality to restart the level yet, but
things like layouts and variables are all set in initialise, and
constants are permanently set in the constructor.

TODO:
~ Right now, everything uses the same cube mesh. This needs to support
custom meshes through use of the mesh manager
~ Polymorphism has been attempted, but has been unsuccessful due to the
lack of pointers used. Will attempt to implement this next. Right now,
only the base block (the wall) class is being used due to overridden
functions not being called.