Level Generation Foundations implemented
Change of plan: Was going to have all the tiles inherit from a base Tile
class where each one can be a bit specific, then use polymporphism to
render etc.

The problem is, polymorphism only works on pointers, so to stop the
tiles from being eradicated, a storage method would have to have been
used to store and point to. A vector is unsuitable, as a re-allocation
will fuck up the pointers. An array would be nice, however due to the
constants used, and the nature of an array, items can't be deleted and
recreated. In our game, the tiles are going to be changed for the
duration of the level, so either we code in the functions to reverse ALL
these changes and implement operators to do so and then create x arrays
for every tile type storing potential tiles.

This is also problematic, as the nature of an array means it can't be
extended to the size of the level, thus let the max amount of tiles be
100 (10*10). This means that, for say 5 types, 500 tiles are always
stored for the sake of using using pointers to cherry pick. This is very
inefficient.

Instead, I have created a tile, and a floortile class (floortiles will
ultimately be their own set, so this is okay. Plus they're rendered on
their own level, this is appropriate).  These classes have an enum
saying what should be spawned on that tile, the floor tile even knows
what's there so it can react if necessary. Then, a simple switch
statement in update, init and render will allow us to adopt different
behaviour, if needed. This may be better or worse, depending on how the
mechanics progress, but for now with basic tiles inheritence isn't
necessary.

Possible future solutions could be a component based extension set for
each tile, using the enum to select which components are needed,
iterating through a component list on update and executing whatevers
there. Again, it depends on the game.

The levels are created inside the levelMGR, which then passes the level
layout to the level class, which handles the spawning and rendering and
interactions of the entire level. The MGR exists only to control the
state of the level.

The LevelMGR class will eventually adopt the singletons pattern