Level-side movement is implemented
The player calls 'move()' on the levelMGR with it's coordinate position,
it's direction in the form of (-1/0/1 x, -1/0/1 y) and a boolean
variable. The boolean will change depending on the success of the move,
and the function returns the new position of the player. If the move
wasn't successful it returns the player's current location.

Underneath is a beast of functionality. Every tile has a new variable
called 'traversable' which essentially is the way to tell if the player
can step on said tile. A check is made on the floor and the the block of
that cell, which both triggers any functionality and returns if it's
doable. If it is doable, moveOff() and moveOn() functions are called on
both the new and previous floor and blocks to trigger more
functionality.

TODO:
- Test if it works - there's conversion from coords to cell and
vice-versa. Don't know if this will even work