Memory leak done, textures fixed, levels made!
- Textures were only applied when the pickup was made, making levels
with no pickups untextured. Fixed.
- Levels are now easily made. Just make sure you ammend the changeLevel
function so that you always call a level that exists, and that it wraps
back to 0 okay.
- Fixed a bug with item pickup - if the end exists earlier than the last
pickup, the check for whether the player has finished occurs before the
level has finished counting. Funnily enough, all tests so far have had
the pickups earlier in the array!
- To fix this bug, I have implemented 'true collected' which takes on
it's little brother's responsiblities. It only updates if the amount
collected is greater than it, meaning that it is always up to date with
the correct value. Fixed!