Added some more levels, fixed end
- All level type enums are shortened to e+5 characters so that making
levels is easier on the eyes to make
- The end condition was met when the end was reached, but it was never
reset if you stepped off. This has been implemented so you have to
finish ON the end tile
- The end tile could be re-stepped on if it was in the falling state.
This has been turned off. If you were to step OFF the end tile you void
the win condition, stopping you from cheesing the game by stepping on a
killed tile.