8th Level is in
- Few more levels are in
- Levels aren't necessarily in a good order, may need restructuring
- The getInfo() function now takes an int as a parameter as some tiles
require an int to access info. The pickup tile also uses said int as a
way of sequencing when it should manifest itself
- In the nextLevel struct, a vector of vectors of ints has been
implemented for "Extra data". In the pickups case, line up the cell with
the pickup's cell and enter how many pickups are required for one to
appear.

TODO:
- MOVEMENT needs a massive fix, as levels are getting more complex my
move hack is getting unreliable. Instead of interpreting the player's
position we should be storing the player's cell and passing that
instead.