Screens are now implemented
- There is now an abstract class called "Screen" which you can inherit
from to make other modes for the game.
- The code from Game has been moved into the GameScreen class, so that
the game we have developed thus far is no longer the only thing that can
be rendered
- In Game, there is now a GameScreen pointer "game" and a Screen pointer
"currentScreen". Use "currentScreen" by pointing it at the screen you
wish to show on screen, this is the trick to having a menu.
- TL;DR - use GameScreen for the game, look at the Game class and the
screens it contains to implement a menuScreen (by inheriting the Screen
class into a new child, MenuScreen perhaps?)