Finalised the the main menu
To work on the menu, you just need to look at the MenuScreen class. The
game starts up in menu mode, and will then will transition to the game
by pressing space.

Use Escape or Q to return to menu and then the same again to quit the
game totally